+++ ArcaneTinmen/Areas/Admin/Controllers/SleevesController.cs
@@ -18,12 +18,14 @@ namespace ArcaneTinmen.Areas.Admin.Controllers
         // GET: Admin/Sleeves
         public ActionResult Index()
         {
+            if (Session["AdminId"] == null) return RedirectToAction("Login", "Management");
             return View(db.Sleeves.ToList());
         }
 
         // GET: Admin/Sleeves/Details/5
         public ActionResult Details(string id)
         {
+            if (Session["AdminId"] == null) return RedirectToAction("Login", "Management");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -39,12 +41,14 @@ namespace ArcaneTinmen.Areas.Admin.Controllers
         // GET: Admin/Sleeves/List
         public ActionResult List()
         {
+            if (Session["AdminId"] == null) return RedirectToAction("Login", "Management");
             return View(db.Sleeves.ToList());
         }
 
         // GET: Admin/Sleeves/Create
         public ActionResult Create()
         {
+            if (Session["AdminId"] == null) return RedirectToAction("Login", "Management");
             return View();
         }
 
@@ -55,6 +59,7 @@ namespace ArcaneTinmen.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "SleeveId,Name,Description,Height,Width,SalePrice,StockAmount,CardImageFileName,BadgeImageFileName")] Sleeve sleeve)
         {
+            if (Session["AdminId"] == null) return RedirectToAction("Login", "Management");
             if (ModelState.IsValid)
             {
                 db.Sleeves.Add(sleeve);
@@ -68,6 +73,7 @@ namespace ArcaneTinmen.Areas.Admin.Controllers
         // GET: Admin/Sleeves/Edit/5
         public ActionResult Edit(string id)
         {
+            if (Session["AdminId"] == null) return RedirectToAction("Login", "Management");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -87,6 +93,7 @@ namespace ArcaneTinmen.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "SleeveId,Name,Description,Height,Width,SalePrice,StockAmount,CardImageFileName,BadgeImageFileName")] Sleeve sleeve)
         {
+            if (Session["AdminId"] == null) return RedirectToAction("Login", "Management");
             if (ModelState.IsValid)
             {
                 db.Entry(sleeve).State = EntityState.Modified;
@@ -99,6 +106,7 @@ namespace ArcaneTinmen.Areas.Admin.Controllers
         // GET: Admin/Sleeves/Delete/5
         public ActionResult Delete(string id)
         {
+            if (Session["AdminId"] == null) return RedirectToAction("Login", "Management");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -116,6 +124,7 @@ namespace ArcaneTinmen.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(string id)
         {
+            if (Session["AdminId"] == null) return RedirectToAction("Login", "Management");
             Sleeve sleeve = db.Sleeves.Find(id);
             db.Sleeves.Remove(sleeve);
             db.SaveChanges();
